refactor(core): migrate pipeline related code to pipelineProcessor service class
