+++ apps/starter/src/index.ts
@@ -26,16 +26,18 @@ import {
 import { ImageGenerationPlugin } from "@maiar-ai/plugin-image";
 import { MCPPlugin } from "@maiar-ai/plugin-mcp";
 import { SearchPlugin } from "@maiar-ai/plugin-search";
-import { TerminalPlugin } from "@maiar-ai/plugin-terminal";
+// import { TerminalPlugin } from "@maiar-ai/plugin-terminal";
 import { TextGenerationPlugin } from "@maiar-ai/plugin-text";
 import { TimePlugin } from "@maiar-ai/plugin-time";
-import {
-  createPostExecutor,
-  periodicPostTrigger,
-  XPlugin
-} from "@maiar-ai/plugin-x";
+
+// import {
+//   createPostExecutor,
+//   periodicPostTrigger,
+//   XPlugin
+// } from "@maiar-ai/plugin-x";
 
 import { SearchPermissionPlugin } from "./lib/plugins/plugin-permissions-search";
+import { TerminalPlugin } from "@maiar-ai/plugin-terminal";
 
 // Suppress deprecation warnings
 process.removeAllListeners("warning");
@@ -67,7 +69,7 @@ async function main() {
     new ImageGenerationPlugin(),
     new TextGenerationPlugin(),
     new TimePlugin(),
-    new SearchPermissionPlugin(["ligma"]),
+    new SearchPermissionPlugin(["0xPBIT"]),
     new SearchPlugin({
       apiKey: process.env.PERPLEXITY_API_KEY as string
     }),
@@ -95,23 +97,23 @@ async function main() {
     ]),
     new CharacterPlugin({
       character: readFileSync(join(process.cwd(), "character.xml"), "utf-8")
-    }),
-    new XPlugin({
-      client_id: process.env.X_CLIENT_ID as string,
-      client_secret: process.env.X_CLIENT_SECRET as string,
-      callback_url: process.env.X_CALLBACK_URL as string,
-      // You can customize which executors and triggers to use
-      // If not specified, all default ones will be used automatically
-      executorFactories: [createPostExecutor],
-      triggerFactories: [periodicPostTrigger]
-    }),
-    new DiscordPlugin({
-      token: process.env.DISCORD_BOT_TOKEN as string,
-      clientId: process.env.DISCORD_CLIENT_ID as string,
-      commandPrefix: "!",
-      executorFactories: [sendMessageExecutor, replyMessageExecutor],
-      triggerFactories: [postListenerTrigger]
     })
+    // new XPlugin({
+    //   client_id: process.env.X_CLIENT_ID as string,
+    //   client_secret: process.env.X_CLIENT_SECRET as string,
+    //   callback_url: process.env.X_CALLBACK_URL as string,
+    //   // You can customize which executors and triggers to use
+    //   // If not specified, all default ones will be used automatically
+    //   executorFactories: [createPostExecutor],
+    //   triggerFactories: [periodicPostTrigger]
+    // }),
+    // new DiscordPlugin({
+    //   token: process.env.DISCORD_BOT_TOKEN as string,
+    //   clientId: process.env.DISCORD_CLIENT_ID as string,
+    //   commandPrefix: "!",
+    //   executorFactories: [sendMessageExecutor, replyMessageExecutor],
+    //   triggerFactories: [postListenerTrigger]
+    // })
   ];
 
   const capabilityAliases: string[][] = [
